Move makeConverter to its own lib
